fix user-defined filters in Command::Github
So the user-defined filters in Command::Github never actually worked
because the `eval` trapped a typo in the JSON fuction name
(`from_from`ne `from_json`). Once we fixed that we discovered that the
JSON doesn't return a hash but instead a hash reference so we have to
properly de-reference that. Rather than leaning on some more abstruse
syntax we add a few more lines of code for clarity and sanity.

Having fixed all of that we then fix the typo in the docs so that people
blindly copying and pasting don't get the wrong thing and spend forever
wondering why.

All of this to prove that `{"assignee":"*"}` will DWIM and give us all
the tickets for a given project overriding the DWIMMEry baked in that
gives us only our own tickets in the common case.